feat: add more git commands

+++ git_base/models/res_user.py
@@ -1,8 +1,11 @@
+import logging
 import os
 import tempfile
 from contextlib import contextmanager
 
 from odoo import fields, models
+
+_logger = logging.getLogger(__name__)
 
 
 class ResUsers(models.Model):
@@ -16,24 +19,26 @@ class ResUsers(models.Model):
     def ssh_env(self):
         """Context manager to set up the SSH environment for Git operations."""
 
-        private_key_file = tempfile.NamedTemporaryFile(delete=False)
-        try:
-            private_key_file.write(self.ssh_private_key.encode("utf-8"))
-            private_key_file.close()
+        if self.ssh_private_key:
+            private_key_file = tempfile.NamedTemporaryFile(delete=False)
+            try:
+                private_key_file.write(self.ssh_private_key.encode("utf-8"))
 
+                # Set the SSH environment variables
+                env = os.environ.copy()
+                env["GIT_SSH_COMMAND"] = f"ssh -i {private_key_file.name}"
+
+                if self.ssh_private_key_password:
+                    env["GIT_ASKPASS"] = "echo"
+                    env["GIT_ASKPASS_STDIN"] = self.ssh_private_key_password
+
+                # Yield the environment to the calling function
+                yield env
+            finally:
+                os.remove(private_key_file.name)
+        else:
             env = os.environ.copy()
-            env[
-                "GIT_SSH_COMMAND"
-            ] = f"ssh -i {private_key_file.name} -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no"
-
-            if self.ssh_private_key_password:
-                env[
-                    "GIT_SSH_COMMAND"
-                ] += " -o PasswordAuthentication=yes -o IdentitiesOnly=yes"
-                env["SSH_ASKPASS"] = "/bin/echo"
-                env["SSH_PASS"] = self.ssh_private_key_password
-
-            # Yield the environment to the calling function
+            env["GIT_SSH_COMMAND"] = ""
+            env["GIT_ASKPASS"] = ""
+            env["GIT_ASKPASS_STDIN"] = ""
             yield env
-        finally:
-            os.remove(private_key_file.name)
